0.9.3 Расширен функционал класса Graph/CentralWidget - 1. Выбор типа выделения (прямоугольная область, горизонтальная область, вертикальная область). 2. Масштабирование осей в ручную (заданием диапазонов для осей). 3. При смене типа графика - масштаб и диапазоны осей остаются на месте!

+++ graph/graphdialog.cpp
@@ -12,7 +12,7 @@ GraphDialog::GraphDialog(Frames *frames, QWidget *parent) :
 //    QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
 
     ui->DataY->addItem("All ");
-    foreach(quint16 value, frames->getClustersLenghtList())
+    foreach(int value, frames->getClustersLenghtList())
         ui->DataY->addItem(QString::number(value));
 
 //    QApplication::restoreOverrideCursor();
@@ -27,9 +27,9 @@ GraphDialog::~GraphDialog()
     delete ui;
 }
 
-quint16 GraphDialog::getCurrentClusterLenght()
+int GraphDialog::getCurrentClusterLenght()
 {
-    return quint16(ui->DataY->currentText().toInt());
+    return ui->DataY->currentText().toInt();
 }
 
 QString GraphDialog::getCurrentX()
